Made it suitable for current usage
