+++ Innov-a-thon/templates/driverprofile.html
@@ -1,0 +1,34 @@
+{% extends "layout1.html" %}
+{% block stuff %}
+  <form method="post" action="{{ url_for('predictdp') }}" enctype="multipart/form-data">
+    {% for text_box in range(3) %}
+<div class="mb-3">
+
+<label for="asensor{{ text_box+1 }}" class="form-label">Enter data Acceleration Sensor {{ text_box+1 }}:</label>
+<input type="number" step="0.0000000000001"  name = "asensor{{ text_box+1 }}" class="form-control" id="asensor{{ text_box+1 }}" placeholder="Enter value" required>
+    <div class="invalid-feedback">
+      Please enter a value
+    </div>
+
+</div>
+
+{% endfor %}
+{% for text_box in range(3) %}
+<div class="mb-3">
+
+<label for="gsensor{{ text_box+1 }}" class="form-label">Enter data for Gyroscope Sensor {{ text_box+1 }}:</label>
+<input type="number" step="0.0000000000001"  name = "gsensor{{ text_box+1 }}" class="form-control" id="gsensor{{ text_box+1 }}" placeholder="Enter value" required>
+    <div class="invalid-feedback">
+      Please enter a value
+    </div>
+
+</div>
+
+{% endfor %}
+
+<button class="btn btn-outline-warning" type="submit">Predict</button>
+ </form>
+    <form method="post" action="{{ url_for('home') }}" enctype="multipart/form-data">
+     <button class="btn btn-outline-light" type="submit">Back to home</button>
+</form>
+{% endblock %}
