+++ Innov-a-thon/templates/pmsm.html
@@ -1,0 +1,119 @@
+{% extends "layout1.html" %}
+{% block stuff %}
+  <form method="post" action="{{ url_for('predictpmsm') }}">
+ 
+<div class="mb-3">
+
+<label for="a0" class="form-label">Enter the Q component of voltage:</label>
+<input type="number" step="0.0000000000001" name = "a0" class="form-control" id="a0" placeholder="Enter value" required>
+    <div class="invalid-feedback">
+      Please enter a value
+    </div>
+
+</div>
+<div class="mb-3">
+
+  <label for="a1" class="form-label">Enter the coolant temperature:</label>
+  <input type="number" step="0.0000000000001" name = "a1" class="form-control" id="a1" placeholder="Enter value" required>
+      <div class="invalid-feedback">
+        Please enter a value
+      </div>
+  
+  </div>
+  <div class="mb-3">
+  <label for="a2" class="form-label">Enter the value of the stator winding:</label>
+  <input type="number" step="0.0000000000001" name = "a2" class="form-control" id="a2" placeholder="Enter value" required>
+      <div class="invalid-feedback">
+        Please enter a value
+      </div>
+  
+  </div>
+
+  <div class="mb-3">
+  <label for="a3" class="form-label">Enter the d component of voltage:</label>
+  <input type="number" step="0.0000000000001"  name = "a3" class="form-control" id="a3" placeholder="Enter value" required>
+      <div class="invalid-feedback">
+        Please enter a value
+      </div>
+  
+  </div>
+
+  <div class="mb-3">
+    <label for="a4" class="form-label">Enter the stator tooth value:</label>
+    <input type="number" step="0.0000000000001" name = "a4" class="form-control" id="a4" placeholder="Enter value" required>
+        <div class="invalid-feedback">
+          Please enter a value
+        </div>
+    
+    </div>
+    <div class="mb-3">
+      <label for="a5" class="form-label">Enter the stator tooth value:</label>
+      <input type="number" step="0.0000000000001" name = "a5" class="form-control" id="a5" placeholder="Enter value" required>
+          <div class="invalid-feedback">
+            Please enter a value
+          </div>
+      
+      </div>
+      <div class="mb-3">
+        <label for="a6" class="form-label">Enter the speed of the motor:</label>
+        <input type="number" step="0.0000000000001"  name = "a6" class="form-control" id="a6" placeholder="Enter value" required>
+            <div class="invalid-feedback">
+              Please enter a value
+            </div>
+        
+        </div>
+
+        <div class="mb-3">
+          <label for="a7" class="form-label">Enter the d component of current:</label>
+          <input type="number" step="0.0000000000001" name = "a7" class="form-control" id="a7" placeholder="Enter value" required>
+              <div class="invalid-feedback">
+                Please enter a value
+              </div>
+          
+          </div>
+
+          <div class="mb-3">
+            <label for="a8" class="form-label">Enter the q component of current:</label>
+            <input type="number" step="0.0000000000001" name = "a8" class="form-control" id="a8" placeholder="Enter value" required>
+                <div class="invalid-feedback">
+                  Please enter a value
+                </div>
+            </div>
+            <div class="mb-3">
+              <label for="a9" class="form-label">Enter the temperature of the stator yoke:</label>
+              <input type="number" step="0.0000000000001" name = "a9" class="form-control" id="a9" placeholder="Enter value" required>
+                  <div class="invalid-feedback">
+                    Please enter a value
+                  </div>
+              
+              </div>
+              <div class="mb-3">
+                <label for="a10" class="form-label">Enter the ambient temperaturee:</label>
+                <input type="number" step="0.0000000000001" name = "a10" class="form-control" id="a10" placeholder="Enter value" required>
+                    <div class="invalid-feedback">
+                      Please enter a value
+                    </div>
+                
+                </div>
+                <div class="mb-3">
+                  <label for="a11" class="form-label">Enter the torque:</label>
+                  <input type="number" step="0.0000000000001"  name = "a11" class="form-control" id="a11" placeholder="Enter value" required>
+                      <div class="invalid-feedback">
+                        Please enter a value
+                      </div>
+                  
+                  </div>
+                  <div class="mb-3">
+                    <label for="a12" class="form-label">Enter the profile id:</label>
+                    <input type="number" step="0.0000000000001" name = "a12" class="form-control" id="a12" placeholder="Enter value" required>
+                        <div class="invalid-feedback">
+                          Please enter a value
+                        </div>
+                    
+                    </div>
+<button class="btn btn-outline-warning" type="submit">Predict</button>
+ </form>
+    <form method="post" action="{{ url_for('home') }}" enctype="multipart/form-data">
+     <button class="btn btn-outline-light" type="submit">Back to home</button>
+</form>
+{% endblock %}
