+++ Innov-a-thon/templates/resultdp.html
@@ -1,0 +1,13 @@
+{%extends "layout1.html"%}
+{% block stuff %}
+<h1 class="dark">Result</h1>
+    <h3 class="dark">The prediction is:</h3>
+    
+        <p>{{ prediction_result[0] }}</p>
+            <form method="post" action="{{ url_for('home') }}" enctype="multipart/form-data">
+
+            <div class="mb-3">
+                <button class="btn btn-outline-info" type="submit">Go to home</button>
+             </div>
+           </form>
+{% endblock %}
