--- Samagam/templates/home.html
@@ -1,13 +1,0 @@
-{% extends "layout1.html" %}
-{% block stuff %}
-
-    <h5 class="dark">Enter driver profiling data: </h5>
-    <form method="post" action="{{ url_for('driverprofiling') }}" enctype="multipart/form-data">
-    <button class="btn btn-outline-success" type="submit">Enter driverprofiling data</button>
-    </form>
-    <h5 class="dark">OR</h5>
-    <form method="post" action="{{ url_for('pmsm') }}" enctype="multipart/form-data">
-    <button class="btn btn-outline-info" type="submit">Enter sensor data</button>
-    </form>
-
-{% endblock %}
