--- Samagam/templates/layout1.html
@@ -1,33 +1,0 @@
-
-<html lang="en" data-bs-theme="dark">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
-    <link rel="stylesheet" type="text/css" href="{{url_for('static',filename='main1.css')}}">
-    
-    <title>Prediction using ML model</title>
-    <style>
-      .dark {
-        background-color: #3e3e42;
-        color: #fff;
-      }
-      .light {
-        background-color: #fff;
-        color: #333;
-      }
-h1 {text-align: center;}
-p {text-align: center;}
-div {text-align: center;}
-    </style>
-  </head>
-  <body class="dark">
-    <legend class="border-bottom mb-4"><h1 class="dark">SmartEV Guardians</h1></legend>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-    <div class="container mt-3">
-      {% block stuff %}{% endblock %}
-    </div>
-   
-
-  </body>
-</html>
