--- Samagam/templates/resultpmsm.html
@@ -1,15 +1,0 @@
-{%extends "layout1.html"%}
-{% block stuff %}
-<h1 class="dark">Result</h1>
-    <h3 class="dark">The predicted value of the permanent magnet temperature is: </h3>
-    
-        <p>{{ prediction_result[0] }}</p>
-    <div>
-    </div>
-            <form method="post" action="{{ url_for('home') }}" enctype="multipart/form-data">
-
-            <div class="mb-3">
-                <button class="btn btn-outline-info" type="submit">Go to home</button>
-             </div>
-           </form>
-{% endblock %}
